pyproject.toml
[tool.poetry]
name = "rabbit-pub-cons"
version = "0.1.0"
description = "Dockerize app for publish and consume message to rabbit"
authors = ["Aleksey Cheremushkin <aleksey.cheremushkin@edgecenter.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
kombu = "^5.4.0"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"
pika = "^1.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
